requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
Django==1.11
django-celery==3.2.2
django-ckeditor==5.6.1
django-js-asset==1.1.0
idna==2.7
kombu==3.0.37
numpy==1.15.0
pandas==0.23.3
paramiko==2.4.1
Pillow==5.2.0
psycopg2==2.7.5
pyasn1==0.4.3
pycparser==2.18
PyMySQL==0.9.2
PyNaCl==1.2.1
pyodbc==4.0.23
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
